Move comments and ratings to offering controller
