+++ eventPlanner/src/main/java/com/ftn/iss/eventPlanner/controller/ServiceController.java
@@ -367,26 +367,4 @@ public class ServiceController {
 	public ResponseEntity<?> delete(@PathVariable("id") int id) {
 		return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 	}
-
-    @PostMapping(value = "{serviceId}/comments/{id}", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<CreatedCommentDTO> createComment(@RequestBody CreateCommentDTO comment, @PathVariable("id") int id) {
-        CreatedCommentDTO createdComment = new CreatedCommentDTO();
-        // id generisan u konstruktoru
-        createdComment.setId(1);
-        createdComment.setContent(comment.getContent());
-        createdComment.setStatus(Status.valueOf("PENDING"));
-        createdComment.setAccountId(id);
-
-        return new ResponseEntity<>(createdComment, HttpStatus.CREATED);
-    }
-    @PostMapping(value = "{serviceId}/ratings/{id}", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<CreatedRatingDTO> createRating(@RequestBody CreateRatingDTO rating, @PathVariable("id") int id) {
-        CreatedRatingDTO createdRating = new CreatedRatingDTO();
-        // id generisan u konstruktoru
-        createdRating.setId(1);
-        createdRating.setScore(rating.getScore());
-        createdRating.setAccountId(id);
-
-        return new ResponseEntity<>(createdRating, HttpStatus.CREATED);
-    }
 }
